Project import generated by Copybara.
GitOrigin-RevId: 640c5b4f7a9e01da7a51933ad140a634cc863712
