+++ arize/pandas/tracing/validation/spans/dataframe_form_validation.py
@@ -1,62 +1,16 @@
-import re
 from datetime import datetime
-from itertools import chain
-from typing import Iterable, List, Optional
+from typing import Iterable, List
 
 import pandas as pd
-from arize.pandas.tracing.columns import (
-    EVAL_COLUMN_PATTERN,
-    EVAL_EXPLANATION_PATTERN,
-    EVAL_LABEL_PATTERN,
-    EVAL_SCORE_PATTERN,
-    SPAN_OPENINFERENCE_COLUMNS,
-    SPAN_OPENINFERENCE_REQUIRED_COLUMNS,
-    SPAN_SPAN_ID_COL,
-    SpanColumnDataType,
-)
+from arize.pandas.tracing.columns import SPAN_OPENINFERENCE_COLUMNS, SpanColumnDataType
 from arize.pandas.tracing.utils import isMissingValue
-from arize.pandas.tracing.validation import errors as tracing_err
-from arize.pandas.validation import errors as err
+from arize.pandas.tracing.validation.common import errors as tracing_err
 from arize.utils.logging import log_a_list, logger
 from arize.utils.types import is_array_of, is_dict_of, is_list_of
 from pandas.api.types import is_bool_dtype, is_numeric_dtype
 
 
-def validate_dataframe_form(
-    spans_dataframe: pd.DataFrame,
-    evals_dataframe: Optional[pd.DataFrame] = None,
-) -> List[err.ValidationError]:
-    _info_dataframe_extra_column_names(spans_dataframe)
-    _info_evals_dataframe_extra_column_names(evals_dataframe)
-    checks = chain(
-        _check_dataframe_index(spans_dataframe),
-        _check_dataframe_required_column_set(
-            spans_dataframe,
-            required_columns=[col.name for col in SPAN_OPENINFERENCE_REQUIRED_COLUMNS],
-        ),
-        _check_dataframe_for_duplicate_columns(spans_dataframe),
-        _check_dataframe_column_content_type(spans_dataframe),
-    )
-    if evals_dataframe is not None:
-        checks = chain(
-            checks,
-            _check_dataframe_index(evals_dataframe),
-            _check_dataframe_required_column_set(
-                evals_dataframe, required_columns=[SPAN_SPAN_ID_COL.name]
-            ),
-            _check_dataframe_for_duplicate_columns(evals_dataframe),
-            _check_evals_dataframe_column_content_type(evals_dataframe),
-        )
-    return list(checks)
-
-
-def _check_dataframe_index(dataframe: pd.DataFrame) -> List[err.InvalidDataFrameIndex]:
-    if (dataframe.index != dataframe.reset_index(drop=True).index).any():
-        return [err.InvalidDataFrameIndex()]
-    return []
-
-
-def _info_dataframe_extra_column_names(
+def _log_info_dataframe_extra_column_names(
     df: pd.DataFrame,
 ) -> None:
     min_col_set = [col.name for col in SPAN_OPENINFERENCE_COLUMNS]
@@ -67,31 +21,6 @@ def _info_dataframe_extra_column_names(
             f"and will be ignored: {log_a_list(list_of_str=extra_col_names, join_word='and')}"
         )
     return None
-
-
-def _check_dataframe_required_column_set(
-    df: pd.DataFrame,
-    required_columns: List[str],
-) -> List[tracing_err.InvalidDataFrameMissingColumns]:
-    existing_columns = set(df.columns)
-    missing_cols = []
-    for col in required_columns:
-        if col not in existing_columns:
-            missing_cols.append(col)
-
-    if missing_cols:
-        return [tracing_err.InvalidDataFrameMissingColumns(missing_cols=missing_cols)]
-    return []
-
-
-def _check_dataframe_for_duplicate_columns(
-    df: pd.DataFrame,
-) -> List[tracing_err.InvalidDataFrameDuplicateColumns]:
-    # Get the duplicated column names from the dataframe
-    duplicate_columns = df.columns[df.columns.duplicated()]
-    if not duplicate_columns.empty:
-        return [tracing_err.InvalidDataFrameDuplicateColumns(duplicate_columns)]
-    return []
 
 
 # TODO(Kiko): Performance improvements
@@ -203,79 +132,6 @@ def _check_dataframe_column_content_type(
         errors.append(
             tracing_err.InvalidDataFrameColumnContentTypes(
                 invalid_type_cols=wrong_string_cols,
-                expected_type="strings",
-            ),
-        )
-    return errors
-
-
-def _info_evals_dataframe_extra_column_names(
-    df: pd.DataFrame,
-) -> None:
-    if df is None:
-        return None
-    irrelevant_columns = [
-        col
-        for col in df.columns
-        if not (pd.Series(col).str.match(EVAL_COLUMN_PATTERN).any() or col == SPAN_SPAN_ID_COL.name)
-    ]
-    if irrelevant_columns:
-        logger.info(
-            "The following columns do not follow the evaluation column naming convention "
-            f"and will be ignored: {log_a_list(list_of_str=irrelevant_columns, join_word='and')}"
-        )
-    return None
-
-
-def _check_evals_dataframe_column_content_type(
-    df: pd.DataFrame,
-) -> List[tracing_err.InvalidDataFrameColumnContentTypes]:
-    wrong_labels_cols = []
-    wrong_scores_cols = []
-    wrong_explanations_cols = []
-    errors = []
-    eval_label_re = re.compile(EVAL_LABEL_PATTERN)
-    eval_score_re = re.compile(EVAL_SCORE_PATTERN)
-    eval_explanation_re = re.compile(EVAL_EXPLANATION_PATTERN)
-    for column in df.columns:
-        if column == SPAN_SPAN_ID_COL.name:
-            if not all(isinstance(value, str) for value in df[column]):
-                errors.append(
-                    tracing_err.InvalidDataFrameColumnContentTypes(
-                        invalid_type_cols=[SPAN_SPAN_ID_COL.name],
-                        expected_type="string",
-                    ),
-                )
-        if eval_label_re.match(column):
-            if not all(isinstance(value, str) or isMissingValue(value) for value in df[column]):
-                wrong_labels_cols.append(column)
-        elif eval_score_re.match(column):
-            if not all(
-                isinstance(value, (int, float)) or isMissingValue(value) for value in df[column]
-            ):
-                wrong_scores_cols.append(column)
-        elif eval_explanation_re.match(column):
-            if not all(isinstance(value, str) or isMissingValue(value) for value in df[column]):
-                wrong_explanations_cols.append(column)
-
-    if wrong_labels_cols:
-        errors.append(
-            tracing_err.InvalidDataFrameColumnContentTypes(
-                invalid_type_cols=wrong_labels_cols,
-                expected_type="strings",
-            ),
-        )
-    if wrong_scores_cols:
-        errors.append(
-            tracing_err.InvalidDataFrameColumnContentTypes(
-                invalid_type_cols=wrong_scores_cols,
-                expected_type="ints or floats",
-            ),
-        )
-    if wrong_explanations_cols:
-        errors.append(
-            tracing_err.InvalidDataFrameColumnContentTypes(
-                invalid_type_cols=wrong_explanations_cols,
                 expected_type="strings",
             ),
         )
