+++ arize/version.py
@@ -1,1 +1,1 @@
-__version__ = "7.15.0"
+__version__ = "7.16.0"
